Partially revert relocation of endpoints
This needs further discussion.
